Bug fix.
git-svn-id: https://svn/repo/rs-lib/trunk@48 12250caf-61d9-4a4f-88eb-f3065243f0ff

+++ rs-csv.h
@@ -181,7 +181,7 @@ extern int rs_csv_valuep (rs_csv_t *__obj);
 extern char *rs_csv_value (rs_csv_t *__obj);
 
 /* Return the length of the most recently parsed field.  */
-extern int rs_csv_length (rs_csv_t *__obj);
+extern size_t rs_csv_length (rs_csv_t *__obj);
 
 RS_CSV_END_DECL
